test(connector): apply review fixes from the two-axis review
Seven confirmed findings from the adversarial review pass, all minor:

- The three surviving proxy/apps/search identity tests now assert
  telemetry privacy through expectTelemetryFreeOfTeamIdentity instead
  of a single not.toHaveProperty("team") check, matching the run test.
- The recording telemetry stub that session.test.ts and
  auth/identity.test.ts each hand-rolled moves to __tests__/helpers.ts
  as createRecordingTelemetry.
- schema-cache.test.ts shares one userScope constant instead of
  repeating the same scope construction fifteen times.
- apps.ts reuses its hasService flag instead of re-evaluating the
  expression.
- Every wire assertion that an identity is absent now checks both
  identity headers; the deleted identity.test.ts only lost the
  x-oo-team-id half of that contract.
- The spec records two implementation-is-right clarifications: the
  five error-rendering CLI tests stay (no catalog-completeness test
  backstops lost messages), and pre-session telemetry recording applies
  only to properties main already recorded before identity resolution.
